Update meson config
Now same flash size as cmake, but executables are not the same.
The reason for flash size was "-fsigned-char" since, I did not know, ARM has default UNSIGNED CHAR, whereas x86 and most textbooks say signed. This flag was probably picked-up at some other repo. Huh.
The reason for diff is not clear, as far as I can tell, most other flags are the same, save for one macro, which is not used anywhere ... it gets added by meson itself ...